+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+
+- Fix: The Tags filter on the Map page wasn't working properly.
+- Update: Improved multiple-selection options and filtering logic in the Filter panel.
+
 ## [4.9.1] - 22.09.2026
 
 - Fix: Import failes if record number column is visible in the table.